feat: 542794 - Turn radio.ts into radio.js
